Add nutrition dashboard

+++ Calories/Model/Nutrition.swift
@@ -12,28 +12,28 @@ let PROTEIN_GOAL_KEY_NAME = "protein_goal"
 let CARBOHYDRATES_GOAL_KEY_NAME = "carbohydrates_goal"
 let FAT_TOTAL_GOAL_KEY_NAME = "fat_total_goal"
 
-let PROTEIN_GOAL_DEFAULT = 60.0
-let CARBOHYDRATES_GOAL_DEFAULT = 250.0
-let FAT_TOTAL_GOAL_DEFAULT = 60.0
+let PROTEIN_GOAL_DEFAULT = 60
+let CARBOHYDRATES_GOAL_DEFAULT = 250
+let FAT_TOTAL_GOAL_DEFAULT = 60
 
 
 // MARK: - Basic Nutrition
 struct BasicNutrition {
     /// Protein
-    let protein: Double
+    let protein: Int
     /// Carbohydrates
-    let carbohydrates: Double
+    let carbohydrates: Int
     /// Total Fat
-    let fatTotal: Double
+    let fatTotal: Int
     
     /// Nutrition goal
     static func goal() -> BasicNutrition {
-        let protein = UserDefaults.standard.double(forKey: PROTEIN_GOAL_KEY_NAME)
-        let carbohydrates = UserDefaults.standard.double(forKey: CARBOHYDRATES_GOAL_KEY_NAME)
-        let fatTotal = UserDefaults.standard.double(forKey: FAT_TOTAL_GOAL_KEY_NAME)
+        let protein = UserDefaults.standard.integer(forKey: PROTEIN_GOAL_KEY_NAME)
+        let carbohydrates = UserDefaults.standard.integer(forKey: CARBOHYDRATES_GOAL_KEY_NAME)
+        let fatTotal = UserDefaults.standard.integer(forKey: FAT_TOTAL_GOAL_KEY_NAME)
         
-        return BasicNutrition(protein: protein != 0.0 ? protein : PROTEIN_GOAL_DEFAULT,
-                              carbohydrates: carbohydrates != 0.0 ? carbohydrates : CARBOHYDRATES_GOAL_DEFAULT,
-                              fatTotal: fatTotal != 0.0 ? fatTotal : FAT_TOTAL_GOAL_DEFAULT)
+        return BasicNutrition(protein: protein != 0 ? protein : PROTEIN_GOAL_DEFAULT,
+                              carbohydrates: carbohydrates != 0 ? carbohydrates : CARBOHYDRATES_GOAL_DEFAULT,
+                              fatTotal: fatTotal != 0 ? fatTotal : FAT_TOTAL_GOAL_DEFAULT)
     }
 }
